// src/app/models/movement/entry/avg-cost-product.ts
export interface AvgCostProductI {
  productId: number;
  productName: string;
  productCategoryId: number;
  productCategory: string;
  productSubcategoryId: number;
  productSubcategory: string;
  productPresentation: string;
  costType: number;
  averageCostValue: number;
  costCurrency: string;
}